+++ src/nodes/index.ts
@@ -13,3 +13,4 @@ export * from "./node-role";
 export * from "./node-exception";
 export * from "./node-system-table";
 export * from "./node-user";
+export * from "./node-index";